It looks like the content generated are not being stored, lets use something like SQL lite to store the data, and later access it or maybe you can suggest something better, don't have to stick to my idea.
and unlike GMB or Marketing and content idea, where we can at least see that it says no optimizations available as in no strategy has been saved yet, on SEO page we don't even see this option so its not even trying to load anything.

and unlike SEO page where we can revise the strategy with AI prompts this option isn't available for GMB, Marketing or Content Ideas.